Rename x0->fixeffs and y0->randeffs
It's somewhat a historical accident we were using x0 and y0 for the fixed and random effects. This rename matches the rename happening in the tutorial to make these values more explicit. Anywhere that they were used in the source was handled too. The function rfx_estimate was renamed to randeffs estimate because it's estimating the values now commonly called randeffs.