Reduce Bayes tolerances and make solver more accurate for value check
The tests used to be set with a lower tolerance than the values were actually generated with, making them only accidentally work (on some computers, on some it failed before!). Thus the tolerance on the solvers was decreased to get a more accurate value, and the tolerance on the test was increased to the vales that it's moving around. It seems like these may be regression tests against previous PuMaS values, in which case they don't make sense because they were not generated with an accurate solver run.